requirements.txt
# CPU-only torch (no CUDA — keeps image small)
--index-url https://download.pytorch.org/whl/cpu
torch
torchaudio

# Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart
httpx>=0.27.0

# STT
faster-whisper>=1.0.0

# TTS
piper-tts>=1.2.0
